requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.1.4
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
idna==3.4
loguru==0.7.0
Markdown==3.2.1
Pillow==10.0.0
pypiwin32==223
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
win32-setctime==1.1.0